+++ lib/rena_rtk.dart
@@ -2,5 +2,6 @@ export 'src/rtk.dart';
 export 'src/rtk_breadcrumb.dart';
 export 'src/rtk_config.dart';
 export 'src/rtk_context.dart';
+export 'src/rtk_device_info.dart';
 export 'src/rtk_error.dart';
 export 'src/rtk_event.dart';